0.9.0 - Make PushMetrics a subclass of Milemarker
As written, the documentation was incorrect - because PushMarker
delegated rather than subclassed, it meant that calling methods that are
part of the public Milemarker interface but that are not defined in
PushMarker did not use the PushMarker implementation. Subclassing rather
than delegating fixes this issue. If in the future, the Milemarker
implementation changes, we may be able to go back to delegation.

This approach uses a clever dynamic approach to subclassing that makes
Pushmetrics a subclass of Milemaker, but also allows us to dynamically
generate Pushmetrics classes as a subclass of other Milemarker classes
or (as in the tests) a mocked version of Milemarker.